c化bug修复 1. ascend_obj.h: task_id数据类型错误 2. database.cpp: HCCLOP表begin字段类型错误 3. host_trace_worker.cpp: 避免生成无用的cannWarehouse对象 4. tree_analyzer.cpp: 修复bug。对ffts+任务在task中存context_id为UINT32_MAX的task；task_id占用采集侧数据结构的低16位 5. tree_builder.cpp: 对于没有kernelEvent的线程依然要通过taskTrack完成建树 6. type_info_db_dumper.cpp: 增加代码可读性
功能增强
1. hash_data.cpp
2. type_data.cpp 出现重复hash 记录warn日志，后续采集侧修订后修改为ERROR级别


# message auto-generated for no-merge-commit merge:
merge new_fix into master

c化bug修复

Created-by: msprofDeveloper

Match-id-0f587ee9f8125f9dc6ae263f1fc24b2b028b2f13

+++ analysis/csrc/entities/ascend_obj.h
@@ -105,7 +105,7 @@ struct HostTask {
     uint64_t taskType = 0;
     uint64_t timeStamp = 0;
     uint32_t streamId = 0;
-    uint32_t taskId = 0;
+    uint16_t taskId = 0;       // 和采集侧的数据类型不一致，采集侧的高低16位会被分别用作batchId和TaskId
     uint32_t contextId = 0;
     uint16_t batchId = 0;
     uint16_t deviceId = 0;
